Create main.py
I created main.py to show the code that runs this color guessing game project. The code is shown on many other GitHub projects. This made me want to put in the code that is shown on main.py right now. I wanted to show everyone how this project works. Before, there was no code that was shown on this GitHub project, and now there is.